# wp-content/plugins/gym-core/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gym Core REST API",
    "description": "REST API for Haanpaa Martial Arts gym management. All endpoints use the gym/v1 namespace and return a standard envelope: { success: true, data: <payload>, meta?: { pagination? } }.",
    "version": "2.1.0",
    "contact": {
      "name": "Andrew Wikel"
    }
  },
  "servers": [
    {
      "url": "/wp-json/gym/v1",
      "description": "WordPress REST API (gym/v1 namespace)"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-WP-Nonce",
        "description": "WordPress REST nonce (cookie authentication)"
      },
      "applicationPassword": {
        "type": "http",
        "scheme": "basic",
        "description": "WordPress Application Password"
      }
    },
...
  },
  "paths": {
...
    "/ranks/promote": {
      "post": {
        "summary": "Promote member to next rank/stripe",
        "tags": ["Rank"],
        "security": [{ "cookieAuth": [] }, { "applicationPassword": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["user_id", "program"],
                "properties": {
                  "user_id": { "type": "integer" },
                  "program": { "type": "string" },
                  "belt": { "type": "string" },
                  "stripes": { "type": "integer" },
                  "notes": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated rank" },
          "403": { "description": "Requires gym_promote_student" }
        }
      }
    },
...
    "/sms/send": {
      "post": {
        "summary": "Send SMS to contact (rate-limited)",
        "tags": ["SMS"],
        "security": [{ "cookieAuth": [] }, { "applicationPassword": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["phone"],
                "properties": {
                  "phone": { "type": "string", "description": "E.164 format" },
                  "message": { "type": "string", "maxLength": 1600 },
                  "template_slug": { "type": "string" },
                  "variables": { "type": "object" },
                  "contact_id": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "SMS sent" },
          "403": { "description": "Requires manage_options or gym_send_sms" },
          "429": { "description": "Rate limit exceeded" }
        }
      }
    },
    "/sms/templates": {
      "get": {
        "summary": "List available SMS templates",
        "tags": ["SMS"],
        "security": [{ "cookieAuth": [] }, { "applicationPassword": [] }],
        "responses": {
          "200": { "description": "SMS templates" }
        }
      }
    }
  },
...
}